requirements.txt
gradio
python-docx
psycopg2
python-dotenv
pywin32 ; sys_platform == 'win32'
docxtpl
requests
msal